requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Deprecated==1.2.14
Flask==3.0.3
Flask-Limiter==3.8.0
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.1
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.2
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
wrapt==1.16.0
